Extracted Mainline mode logic into it's own class
